extract find_items_with_different_depths method
